client-test/pyproject.toml
[project]
name = "client-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.16.1",
    "pika>=1.3.2",
    "pydantic>=2.11.5",
    "python-dotenv>=1.1.0",
    "requests>=2.32.4",
    "sqlalchemy>=2.0.41",
]
